requirements.txt
streamlit
google-generativeai
PyPDF2
python-docx
pandas
plotly
numpy
xlsxwriter
difflib